Feature

+++ database/migrations/2020_01_28_000001_create_debit_card_transactions_table.php
@@ -13,20 +13,22 @@ class CreateDebitCardTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('debit_card_transactions', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('debit_card_id');
-            $table->integer('amount');
-            $table->string('currency_code');
-            $table->timestamps();
-            $table->softDeletes();
-
-            $table->foreign('debit_card_id')
-                ->references('id')
-                ->on('debit_cards')
-                ->onUpdate('cascade')
-                ->onDelete('restrict');
-        });
+        if (!Schema::hasTable("debit_card_transactions")) {
+            Schema::create('debit_card_transactions', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('debit_card_id');
+                $table->integer('amount');
+                $table->string('currency_code');
+                $table->timestamps();
+                $table->softDeletes();
+    
+                $table->foreign('debit_card_id')
+                    ->references('id')
+                    ->on('debit_cards')
+                    ->onUpdate('cascade')
+                    ->onDelete('restrict');
+            });
+        }
     }
 
     /**
